requirements.txt
customtkinter==0.3
darkdetect==0.8.0
mysql-connector==2.2.9
numpy==1.26.1
pandas==2.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
